schedule to once a month
I thought of stopping it but it's still kinda interesting
to look at changes the used technologies, webtrackers, cookies,
http-headers, etc.. Let it run once a month. That way it can
run for a couple of years before hitting the 1Gb free github
limit.

Nevertheless, thank you github :thumbsup:

It's just great to be able to do that